requirements.txt
Flask==2.2.5
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.1.1
Werkzeug==2.2.3
python-telegram-bot==20.7
python-binance==1.0.19
pandas==2.1.4
numpy==1.26.2
requests==2.31.0
python-dotenv==1.0.0
TA-Lib==0.6.3 